requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
Django==4.0.2
djangorestframework==3.13.1
lxml==4.7.1
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
XlsxWriter==3.0.2
